camera corrupted if scene not persistent - breaking the build. - FIXED. missing data files added.
git-svn-id: http://projects.hashstash.in/svn/trivial@28 659a0b78-bc15-4dc9-aeea-a5f01ff3c5d0

+++ src/TrivialCamera.h
@@ -86,8 +86,8 @@ protected:
     //Scene* _scene;
 
 /* half your own width */
-    float _halfWidth;
-    float _halfHeight;
+    int _halfWidth;
+    int _halfHeight;
 
     bool _ready;
